requirement.txt
flask-restful
mt5linux
python-dotenv
sqlalchemy
pymysql
pyyaml